Fix App initial page load.
While this worked previously, the current build and use of Android
Webview no longer refreshed the table with the default (or saved state,
if any). restoreUIState was never called. Looks like Webview either
overrides or drops $(window).on("load") callbacks.

Since we only need DOM loaded, can restorUIState in jquery .ready()
instead of window.on("load") and that fixes the Webview problem.
Page is refreshed on app start, and it works on desktop browsers too.